@@ -1,13 +1,17 @@
 import * as React from "react";
-import { useState } from 'react';
-import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import { useState, useEffect } from 'react';
+import { BrowserRouter as Router, Route, Routes, Navigate, Outlet } from 'react-router-dom';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import ProtectedRoute from './components/ProtectedRoute/ProtectedRoute';
+import 'intro.js/introjs.css';
+import { AuthProvider } from './contexts/AuthContext';
 
 import LogIn from "./components/Auth/LogIn/LogIn";
 import SignUp from "./components/Auth/SignUp/SignUp";
 import OTPVerification from './components/Auth/OTPVerification/OTPVerification';
+import ForgotPassword from './components/Auth/ForgotPassword/ForgotPassword';
+import ResetPassword from './components/Auth/ResetPassword/ResetPassword';
 
 import Sidebar from './components/Sidebar/Sidebar';
 import Dashboard from './pages/Dashboard/Dashboard'
@@ -15,13 +19,14 @@ import PrivateMemos from './pages/PrivateMemos/PrivateMemos';
 import HealthReport from "./pages/HealthReport/HealthReport";
 
 import PollutionTracker from './pages/PollutionTracker/PollutionTracker';
-import Calories from "./pages/Calories/Calories";
-import Profile from "./pages/Profile/Profile";
 import HealthTips from "./pages/HealthTips/HealthTips";
-import ExerciseRoutines from "./pages/ExerciseRoutines/ExerciseRoutines";
-import WellnessHub from "./pages/WellnessHub/WellnessHub";
 import EmergencyContacts from "./pages/EmergencyContacts/EmergencyContacts";
 
+import Calories from "./pages/Calories/Calories";
+import ExerciseRoutines from "./pages/ExerciseRoutines/ExerciseRoutines";
+import WellnessHub from "./pages/WellnessHub/WellnessHub";
+
+import Profile from "./pages/Profile/Profile";
 import Settings from "./pages/Settings/Settings";
 import Help from './pages/Help/Help';
 import TermsOfUse from './pages/TermsOfUse/TermsOfUse';
@@ -35,16 +40,22 @@ import './App.css';
 function App() {
     const [isDarkMode, setIsDarkMode] = useState(() => {
         const savedTheme = localStorage.getItem('theme');
-        return savedTheme ? savedTheme === 'dark' : false;
+        return savedTheme === 'dark';
     });
 
-    const toggleTheme = () => {
-        const newTheme = !isDarkMode;
-        setIsDarkMode(newTheme);
-        localStorage.setItem('theme', newTheme ? 'dark' : 'light');
+    useEffect(() => {
+        localStorage.setItem('theme', isDarkMode ? 'dark' : 'light');
+        document.body.classList.toggle('dark-mode', isDarkMode);
+    }, [isDarkMode]);
+
+    const isAuthenticated = () => {
+        const token = localStorage.getItem('token');
+        const userId = localStorage.getItem('userId');
+        return !!(token && userId);
     };
 
-    const handleAuthSuccess = () => {
+    const toggleTheme = () => {
+        setIsDarkMode(prevMode => !prevMode);
     };
 
     const AppLayout = ({ children }) => (
@@ -57,86 +68,62 @@ function App() {
     );
 
     return (
-        <>
-            <Router>
-                <Routes>
-                    {/* Public Routes */}
-                    <Route path="/" element={<SignUp isDarkMode={isDarkMode} toggleTheme={toggleTheme} onAuthSuccess={handleAuthSuccess} />} />
-                    <Route path="/log-in" element={<LogIn isDarkMode={isDarkMode} toggleTheme={toggleTheme} onAuthSuccess={handleAuthSuccess} />} />
-                    <Route path="/verify-otp" element={<OTPVerification isDarkMode={isDarkMode} toggleTheme={toggleTheme} />} />
+        <Router>
+            <AuthProvider>
+                <div className={isDarkMode ? 'dark' : 'light'}>
+                    <ToastContainer
+                        position="top-right"
+                        theme={isDarkMode ? 'dark' : 'light'}
+                    />
+                    <Routes>
+                        {/* Public routes */}
+                        <Route 
+                            path="/" 
+                            element={
+                                isAuthenticated() ? (
+                                    <Navigate to="/dashboard" replace />
+                                ) : (
+                                    <SignUp isDarkMode={isDarkMode} toggleTheme={toggleTheme} />
+                                )
+                            } 
+                        />
+                        <Route path="/log-in" element={<LogIn isDarkMode={isDarkMode} toggleTheme={toggleTheme} />} />
+                        <Route path="/verify-otp" element={<OTPVerification isDarkMode={isDarkMode} toggleTheme={toggleTheme} />} />
+                        <Route path="/forgot-password" element={<ForgotPassword isDarkMode={isDarkMode} toggleTheme={toggleTheme} />} />
+                        <Route path="/reset-password" element={<ResetPassword isDarkMode={isDarkMode} toggleTheme={toggleTheme} />} />
 
-                    {/* Protected Routes */}
-                    <Route path="/dashboard" element={
-                        <ProtectedRoute>
-                            <AppLayout>
-                                <Dashboard isDarkMode={isDarkMode} />
-                            </AppLayout>
-                        </ProtectedRoute>
-                    } />
-                    
-                    <Route path="/sticky-notes" element={
-                        <ProtectedRoute>
-                            <AppLayout><PrivateMemos isDarkMode={isDarkMode} /></AppLayout>
-                        </ProtectedRoute>
-                    } />
-                    <Route path="/health-report" element={
-                        <ProtectedRoute>
-                            <AppLayout><HealthReport isDarkMode={isDarkMode} /></AppLayout>
-                        </ProtectedRoute>
-                    } />
-                    <Route path="/bmr-calculator" element={
-                            <AppLayout><Calories isDarkMode={isDarkMode} /></AppLayout>
-                    } />
-                    <Route path="/pollution-tracker" element={
-                        <ProtectedRoute>
-                            <AppLayout><PollutionTracker isDarkMode={isDarkMode} /></AppLayout>
-                        </ProtectedRoute>
-                    } />
-                    <Route path="/health-tips" element={
-                            <AppLayout><HealthTips isDarkMode={isDarkMode} /></AppLayout>
-                    } />
-                    <Route path="/exercise-routines" element={
-                            <AppLayout><ExerciseRoutines isDarkMode={isDarkMode} /></AppLayout>
-                    } />
-                    <Route path="/wellness-hub" element={
-                            <AppLayout><WellnessHub isDarkMode={isDarkMode} /></AppLayout>
-                    } />
-                    <Route path="/emergency-contacts" element={
-                        <ProtectedRoute>
-                            <AppLayout><EmergencyContacts isDarkMode={isDarkMode} /></AppLayout>
-                        </ProtectedRoute>
-                    } />
-                    <Route path="/profile" element={
-                        <ProtectedRoute>
-                            <AppLayout><Profile isDarkMode={isDarkMode} /></AppLayout>
-                        </ProtectedRoute>
-                    } />
-                    <Route path="/settings" element={
-                        <ProtectedRoute>
-                            <AppLayout><Settings isDarkMode={isDarkMode} /></AppLayout>
-                        </ProtectedRoute>
-                    } />
-                    <Route path="/help" element={
-                            <AppLayout><Help isDarkMode={isDarkMode} /></AppLayout>
-                    } />
-                    <Route path="/terms-of-use" element={
-                            <AppLayout><TermsOfUse isDarkMode={isDarkMode} /></AppLayout>
-                    } />
-                    <Route path="/privacy-policy" element={
-                            <AppLayout><PrivacyPolicy isDarkMode={isDarkMode} /></AppLayout>
-                    } />
-                    <Route path="/analytics" element={
-                        <ProtectedRoute>
-                            <AppLayout><Analytics isDarkMode={isDarkMode} /></AppLayout>
-                        </ProtectedRoute>
-                    } />
+                        {/* Protected routes */}
+                        <Route
+                            element={
+                                <ProtectedRoute>
+                                    <AppLayout isDarkMode={isDarkMode}>
+                                        <Outlet /> {/* This will render the child routes */}
+                                    </AppLayout>
+                                </ProtectedRoute>
+                            }
+                        >
+                            <Route path="/dashboard" element={<Dashboard isDarkMode={isDarkMode} />} />
+                            <Route path="/sticky-notes" element={<PrivateMemos isDarkMode={isDarkMode} />} />
+                            <Route path="/health-report" element={<HealthReport isDarkMode={isDarkMode} />} />
+                            <Route path="/bmr-calculator" element={<Calories isDarkMode={isDarkMode} />} />
+                            <Route path="/pollution-tracker" element={<PollutionTracker isDarkMode={isDarkMode} />} />
+                            <Route path="/health-tips" element={<HealthTips isDarkMode={isDarkMode} />} />
+                            <Route path="/exercise-routines" element={<ExerciseRoutines isDarkMode={isDarkMode} />} />
+                            <Route path="/wellness-hub" element={<WellnessHub isDarkMode={isDarkMode} />} />
+                            <Route path="/emergency-contacts" element={<EmergencyContacts isDarkMode={isDarkMode} />} />
+                            <Route path="/profile" element={<Profile isDarkMode={isDarkMode} />} />
+                            <Route path="/settings" element={<Settings isDarkMode={isDarkMode} />} />
+                            <Route path="/help" element={<Help isDarkMode={isDarkMode} />} />
+                            <Route path="/terms-of-use" element={<TermsOfUse isDarkMode={isDarkMode} />} />
+                            <Route path="/privacy-policy" element={<PrivacyPolicy isDarkMode={isDarkMode} />} />
+                            <Route path="/analytics" element={<Analytics isDarkMode={isDarkMode} />} />
+                        </Route>
 
-                    {/* 404 Route*/}
-                    <Route path="*" element={<NotFound isDarkMode={isDarkMode} />} />
-                </Routes>
-            </Router>
-            <ToastContainer position="top-right" theme={isDarkMode ? 'dark' : 'light'} />
-        </>
+                        <Route path="*" element={<NotFound isDarkMode={isDarkMode} />} />
+                    </Routes>
+                </div>
+            </AuthProvider>
+        </Router>
     );
 }
 
