@@ -1,5 +1,5 @@
 import React from 'react';
-import { motion, AnimatePresence } from 'framer-motion';
+import { motion } from 'framer-motion';
 import './BreathingCircle.css';
 import { FaTimes } from 'react-icons/fa';
 
