@@ -1,61 +1,13 @@
 import React, { useState, useEffect } from 'react';
 import { Map, NavigationControl, Marker, Source, Layer } from 'react-map-gl';
-import './PollutionTracker.css';
 import { FaMapMarkerAlt, FaInfoCircle } from 'react-icons/fa';
 import mapboxgl from 'mapbox-gl';
+import { pollutionZones } from '../../data/pollution-data';
+import { legendItems } from '../../data/legend-items';
+import { INITIAL_VIEW_STATE } from '../../data/initial-view-state';
+import './PollutionTracker.css';
 
 const MAPBOX_API_KEY = import.meta.env.VITE_MAPBOX_API_KEY;
-
-const INITIAL_VIEW_STATE = {
-  latitude: 40.6935,
-  longitude: -73.9866,
-  zoom: 13,
-  bearing: 0,
-  pitch: 0,
-};
-
-// Mock data for pollution zones
-const pollutionZones = [
-  {
-    id: '1',
-    coordinates: [40.6935, -73.9866],
-    level: 'high',
-    radius: 500,
-    data: {
-      aqi: 156,
-      pm25: 75.2,
-      pm10: 142.8,
-    }
-  },
-  {
-    id: '2',
-    coordinates: [40.6895, -73.9845],
-    level: 'medium',
-    radius: 300,
-    data: {
-      aqi: 89,
-      pm25: 35.5,
-      pm10: 82.3,
-    }
-  },
-  {
-    id: '3',
-    coordinates: [40.6915, -73.9825],
-    level: 'low',
-    radius: 400,
-    data: {
-      aqi: 42,
-      pm25: 12.8,
-      pm10: 38.5,
-    }
-  },
-];
-
-const legendItems = [
-  { color: '#FF4444', label: 'High Pollution', description: 'AQI > 150' },
-  { color: '#FFB344', label: 'Medium Pollution', description: 'AQI 51-150' },
-  { color: '#4CAF50', label: 'Low Pollution', description: 'AQI 0-50' },
-];
 
 const Legend = () => (
   <div className="legend">
@@ -132,20 +84,41 @@ export default function PollutionTracker({ isDarkMode }) {
     }
   }, []);
 
-  const geojsonData = {
-    type: 'FeatureCollection',
-    features: pollutionZones.map(zone => ({
-      type: 'Feature',
-      geometry: {
-        type: 'Point',
-        coordinates: [zone.coordinates[1], zone.coordinates[0]]
-      },
-      properties: {
-        id: zone.id,
-        level: zone.level,
-        radius: zone.radius
-      }
-    }))
+  const renderPollutionZones = () => {
+    return pollutionZones.map(zone => (
+      <Source
+        key={zone.id}
+        type="geojson"
+        data={{
+          type: 'Feature',
+          geometry: {
+            type: 'Point',
+            coordinates: [zone.coordinates[1], zone.coordinates[0]]
+          },
+          properties: {
+            level: zone.level
+          }
+        }}
+      >
+        <Layer
+          id={`zone-${zone.id}`}
+          type="circle"
+          paint={{
+            'circle-radius': [
+              'interpolate',
+              ['linear'],
+              ['zoom'],
+              10, zone.radius / 100,
+              15, zone.radius / 50
+            ],
+            'circle-color': getPollutionColor(zone.level),
+            'circle-opacity': 0.6,
+            'circle-blur': 0.5,
+          }}
+          onClick={() => setSelectedZone(zone)}
+        />
+      </Source>
+    ));
   };
 
   return (
@@ -157,39 +130,12 @@ export default function PollutionTracker({ isDarkMode }) {
           style={{ width: '100%', height: '100%' }}
           mapStyle={isDarkMode ? "mapbox://styles/mapbox/dark-v10" : "mapbox://styles/mapbox/light-v10"}
           mapboxAccessToken={MAPBOX_API_KEY}
+          onClick={() => setSelectedZone(null)}
         >
           <NavigationControl position="top-right" />
           
-          {/* Render pollution zones */}
-          {pollutionZones.map(zone => (
-            <Source
-              key={zone.id}
-              type="geojson"
-              data={{
-                type: 'Feature',
-                geometry: {
-                  type: 'Point',
-                  coordinates: [zone.coordinates[1], zone.coordinates[0]]
-                },
-                properties: {
-                  level: zone.level
-                }
-              }}
-            >
-              <Layer
-                type="circle"
-                paint={{
-                  'circle-radius': zone.radius / 50,
-                  'circle-color': getPollutionColor(zone.level),
-                  'circle-opacity': 0.6,
-                  'circle-blur': 0.5,
-                }}
-                onClick={() => setSelectedZone(zone)}
-              />
-            </Source>
-          ))}
+          {renderPollutionZones()}
 
-          {/* User location marker */}
           {userLocation && (
             <Marker
               longitude={userLocation.longitude}
