@@ -1,19 +1,53 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { motion } from 'framer-motion';
-import { FaArrowLeft, FaMoon, FaSun, FaBell, FaUser, FaEnvelope, FaClock, FaRuler } from 'react-icons/fa';
+import { FaMoon, FaSun, FaBell, FaUser, FaEnvelope, FaRuler } from 'react-icons/fa';
 import { IoMdNotifications, IoMdNotificationsOff } from 'react-icons/io';
 import { toast } from 'react-toastify';
+import { fetchWithAuth, API_ENDPOINTS } from '../../utils/api';
+import Spinner from '../../components/Spinner/Spinner';
 
 const SettingsPage = ({ isDarkMode, toggleDarkMode }) => {
   const [settings, setSettings] = useState({
     notifications: true,
-    email: 'evansachie01@gmail.com',
-    username: 'Evans',
+    email: '',
+    username: '',
     units: 'metric',
     language: 'english',
     workoutReminders: true,
     emergencyContacts: true
   });
+  const [isLoading, setIsLoading] = useState(true);
+
+  useEffect(() => {
+    const fetchUserData = async () => {
+      setIsLoading(true);
+      try {
+        const response = await fetchWithAuth(`${API_ENDPOINTS.GET_USER}?id=${localStorage.getItem('userId')}`, {
+          method: 'GET',
+        });
+        
+        setSettings(prev => ({
+          ...prev,
+          email: response.email,
+          username: response.userName
+        }));
+      } catch (error) {
+        toast.error('Failed to fetch user data');
+        console.error('Error fetching user data:', error);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
+    fetchUserData();
+  }, []);
+
+  const handleThemeToggle = () => {
+    toggleDarkMode();
+    // Save theme preference to localStorage
+    localStorage.setItem('theme', !isDarkMode ? 'dark' : 'light');
+    toast.success(`Switched to ${!isDarkMode ? 'dark' : 'light'} mode`);
+  };
 
   const handleSettingChange = (key, value) => {
     setSettings(prev => ({
@@ -26,13 +60,12 @@ const SettingsPage = ({ isDarkMode, toggleDarkMode }) => {
     toast.success('Settings updated successfully!');
   };
 
-
   const SettingSection = ({ title, children }) => (
     <motion.div
       initial={{ opacity: 0, y: 20 }}
       animate={{ opacity: 1, y: 0 }}
       className={`p-6 rounded-lg shadow-lg mb-6 ${
-        isDarkMode ? 'bg-gray-800' : 'bg-white'
+        isDarkMode ? 'bg-[#2d2d2d]' : 'bg-white'
       }`}
     >
       <h2 className="text-xl font-semibold mb-4">{title}</h2>
@@ -63,21 +96,24 @@ const SettingsPage = ({ isDarkMode, toggleDarkMode }) => {
     <div className={`min-h-screen p-6 ${isDarkMode ? 'dark-mode' : 'bg-gray-50 text-gray-900'}`}>
       {/* Header */}
       <div className="flex items-center gap-4 mb-8">
-        <h1 className="text-2xl font-bold">Settings</h1>
+        <h1 className="text-3xl font-bold mb-4">Settings</h1>
       </div>
 
       {/* Theme Settings */}
       <SettingSection title="Appearance">
         <div className="flex items-center justify-between">
           <div className="flex items-center gap-3">
-            {isDarkMode ? <FaMoon className="text-xl" /> : <FaSun className="text-xl text-yellow-500" />}
+            {isDarkMode ? <FaMoon className="text-xl text-yellow-500" /> : <FaSun className="text-xl text-yellow-500" />}
             <span>Dark Mode</span>
           </div>
-          <ToggleSwitch enabled={isDarkMode} onChange={toggleDarkMode} />
+          <ToggleSwitch
+            enabled={isDarkMode}
+            onChange={handleThemeToggle}
+          />
         </div>
       </SettingSection>
 
-      {/* Profile Settings */}
+      {/* Profile Section */}
       <SettingSection title="Profile">
         <div className="space-y-4">
           <div className="flex items-center gap-3 mb-4">
@@ -86,16 +122,22 @@ const SettingsPage = ({ isDarkMode, toggleDarkMode }) => {
               <div className={`block text-sm mb-1 ${isDarkMode ? 'text-gray-300' : 'text-gray-600'}`}>
                 Username
               </div>
-              <input
-                type="text"
-                value={settings.username}
-                onChange={(e) => handleSettingChange('username', e.target.value)}
-                className={`w-full p-2 rounded border ${
-                  isDarkMode 
-                    ? 'bg-gray-700 border-gray-600' 
-                    : 'bg-white border-gray-300'
-                }`}
-              />
+              {isLoading ? (
+                <div className="flex justify-center py-2">
+                  <Spinner size="small" color={isDarkMode ? '#4285F4' : '#4285F4'} />
+                </div>
+              ) : (
+                <input
+                  type="text"
+                  value={settings.username}
+                  readOnly
+                  className={`w-full p-2 rounded border ${
+                    isDarkMode 
+                      ? 'bg-gray-700 border-gray-600' 
+                      : 'bg-gray-200 border-gray-300'
+                  } cursor-not-allowed`}
+                />
+              )}
             </div>
           </div>
           <div className="flex items-center gap-3">
@@ -104,16 +146,22 @@ const SettingsPage = ({ isDarkMode, toggleDarkMode }) => {
               <div className={`block text-sm mb-1 ${isDarkMode ? 'text-gray-300' : 'text-gray-600'}`}>
                 Email
               </div>
-              <input
-                type="email"
-                value={settings.email}
-                onChange={(e) => handleSettingChange('email', e.target.value)}
-                className={`w-full p-2 rounded border ${
-                  isDarkMode 
-                    ? 'bg-gray-700 border-gray-600' 
-                    : 'bg-white border-gray-300'
-                }`}
-              />
+              {isLoading ? (
+                <div className="flex justify-center py-2">
+                  <Spinner size="small" color={isDarkMode ? '#4285F4' : '#4285F4'} />
+                </div>
+              ) : (
+                <input
+                  type="email"
+                  value={settings.email}
+                  readOnly
+                  className={`w-full p-2 rounded border ${
+                    isDarkMode 
+                      ? 'bg-gray-700 border-gray-600' 
+                      : 'bg-gray-200 border-gray-300'
+                  } cursor-not-allowed`}
+                />
+              )}
             </div>
           </div>
         </div>
